# contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ExperienceLevel": {
        "enum": [
          "ANY",
          "ENTRY",
          "JUNIOR",
          "MID",
          "SENIOR",
          "LEAD",
          "EXECUTIVE"
        ],
        "title": "ExperienceLevel",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchCreate": {
        "properties": {
          "employment_types": {
            "items": {
              "enum": [
                "FULL_TIME",
                "PART_TIME",
                "CONTRACT",
                "TEMPORARY",
                "INTERNSHIP",
                "FREELANCE"
              ],
              "type": "string"
            },
            "title": "Employment Types",
            "type": "array"
          },
          "excluded_terms": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Excluded Terms",
            "type": "array"
          },
          "experience_level": {
            "$ref": "#/components/schemas/ExperienceLevel",
            "default": "ANY"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "posting_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 30,
            "title": "Posting Age Days"
          },
          "raw_intent": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Intent"
          },
          "required_terms": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Required Terms",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/WatchSourceInput"
            },
            "maxItems": 20,
            "title": "Sources",
            "type": "array"
          },
          "target_terms": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Target Terms",
            "type": "array"
          },
          "work_arrangements": {
            "items": {
              "enum": [
                "ON_SITE",
                "HYBRID",
                "REMOTE"
              ],
              "type": "string"
            },
            "title": "Work Arrangements",
            "type": "array"
          }
        },
        "required": [
          "target_terms"
        ],
        "title": "WatchCreate",
        "type": "object"
      },
      "WatchRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "employment_types": {
            "items": {
              "type": "string"
            },
            "title": "Employment Types",
            "type": "array"
          },
          "excluded_terms": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Terms",
            "type": "array"
          },
          "experience_level": {
            "title": "Experience Level",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "posting_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Age Days"
          },
          "raw_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Intent"
          },
          "required_terms": {
            "items": {
              "type": "string"
            },
            "title": "Required Terms",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/WatchSourceRead"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/WatchStatus"
          },
          "target_terms": {
            "items": {
              "type": "string"
            },
            "title": "Target Terms",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "work_arrangements": {
            "items": {
              "type": "string"
            },
            "title": "Work Arrangements",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "target_terms",
          "required_terms",
          "excluded_terms",
          "locations",
          "work_arrangements",
          "employment_types",
          "experience_level",
          "raw_intent",
          "posting_age_days",
          "sources",
          "created_at",
          "updated_at"
        ],
        "title": "WatchRead",
        "type": "object"
      },
      "WatchRunRead": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discovered_count": {
            "title": "Discovered Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "matched_count": {
            "title": "Matched Count",
            "type": "integer"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "sources_failed": {
            "title": "Sources Failed",
            "type": "integer"
          },
          "sources_succeeded": {
            "title": "Sources Succeeded",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "watch_id": {
            "title": "Watch Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "watch_id",
          "status",
          "trigger",
          "correlation_id",
          "created_at",
          "completed_at",
          "outcome",
          "sources_succeeded",
          "sources_failed",
          "discovered_count",
          "matched_count"
        ],
        "title": "WatchRunRead",
        "type": "object"
      },
      "WatchSourceInput": {
        "properties": {
          "adapter_key": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[a-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Key"
          },
          "platform_key": {
            "anyOf": [
              {
                "maxLength": 64,
                "pattern": "^[a-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Key"
          },
          "source_kind": {
            "enum": [
              "PLATFORM",
              "CUSTOM_URL"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "source_kind"
        ],
        "title": "WatchSourceInput",
        "type": "object"
      },
      "WatchSourceRead": {
        "properties": {
          "adapter_key": {
            "title": "Adapter Key",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "platform_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Key"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "source_kind",
          "platform_key",
          "adapter_key",
          "url"
        ],
        "title": "WatchSourceRead",
        "type": "object"
      },
      "WatchStatus": {
        "enum": [
          "DRAFT",
          "ACTIVE",
          "PAUSED",
          "ARCHIVED"
        ],
        "title": "WatchStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "DireHire API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/watches": {
      "get": {
        "operationId": "list_watches_api_v1_watches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchRead"
                  },
                  "title": "Response List Watches Api V1 Watches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Watches",
        "tags": [
          "Job Watches"
        ]
      },
      "post": {
        "operationId": "create_watch_api_v1_watches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watch",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/platforms": {
      "get": {
        "operationId": "list_platforms_api_v1_watches_platforms_get",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Platforms Api V1 Watches Platforms Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Platforms",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}": {
      "delete": {
        "operationId": "delete_watch_api_v1_watches__watch_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watch",
        "tags": [
          "Job Watches"
        ]
      },
      "get": {
        "operationId": "get_watch_api_v1_watches__watch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watch",
        "tags": [
          "Job Watches"
        ]
      },
      "put": {
        "operationId": "replace_watch_api_v1_watches__watch_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Watch",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}/activate": {
      "post": {
        "operationId": "activate_watch_api_v1_watches__watch_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Watch",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}/archive": {
      "post": {
        "operationId": "archive_watch_api_v1_watches__watch_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Watch",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}/pause": {
      "post": {
        "operationId": "pause_watch_api_v1_watches__watch_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Watch",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}/runs": {
      "get": {
        "operationId": "list_watch_runs_api_v1_watches__watch_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchRunRead"
                  },
                  "title": "Response List Watch Runs Api V1 Watches  Watch Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Watch Runs",
        "tags": [
          "Job Watches"
        ]
      },
      "post": {
        "operationId": "run_watch_api_v1_watches__watch_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Watch",
        "tags": [
          "Job Watches"
        ]
      }
    },
    "/api/v1/watches/{watch_id}/sources/import-csv": {
      "post": {
        "operationId": "import_watch_sources_api_v1_watches__watch_id__sources_import_csv_post",
        "parameters": [
          {
            "in": "path",
            "name": "watch_id",
            "required": true,
            "schema": {
              "title": "Watch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Watch Sources",
        "tags": [
          "Job Watches"
        ]
      }
    }
  }
}